requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
drf-yasg==1.21.7
Pillow==10.3.0
celery==5.4.0
redis==5.0.6
django-filter==24.2
